+++ src/renderer/src/lib/query-client.ts
@@ -6,8 +6,10 @@ const queryClient = new QueryClient({
       refetchOnWindowFocus: false,
       refetchIntervalInBackground: false,
       refetchOnReconnect: false,
-      gcTime: 0,
-      staleTime: 0,
+      // 10 分钟后回收缓存
+      gcTime: 10 * 60 * 1000,
+      // 5 分钟内视为新鲜，不重新请求
+      staleTime: 5 * 60 * 1000,
     },
   },
 })