+++ src/renderer/src/request/index.ts
@@ -1,7 +1,6 @@
 import { bangumi } from './api/bangumi'
 import { comment } from './api/comment'
 import { match } from './api/match'
-import { related } from './api/related'
 import { search } from './api/search'
 
 export const apiClient = {
@@ -9,5 +8,4 @@ export const apiClient = {
   comment,
   search,
   bangumi,
-  related,
 }